+++ examples/js/_test.js
@@ -1,4 +1,5 @@
-import { Fibonacci } from "../../src/index.js";
+import { Fibonacci, randomRGBAColor } from "../../src/index.js";
 // import { allCharactersSame } from "../../../src/allCharactersSame.js";
 
 console.log(new Fibonacci());
+console.log(randomRGBAColor());