+++ CHANGELOG.md
@@ -18,15 +18,16 @@ This file was based on [this template](https://gist.github.com/juampynr/4c18214a
  ### Fixed
 - randomInt - Fix return a number between the given min and max;
 - arrayChoice - Fixed not returning the last element;
+- isObjectEmpty - Check is the given object is empty;
+
  #### Breaking Changes
  - rockPaperScissor - move to /Games #60;
 
- 
 ## [1.2.7] - 15-07-2022
 ### Added
  - [Full Screen](https://gitlab.com/201flaviosilva/utilsjs/-/commit/cedd102ff36ffd712d0d20be30f21b9003253027) #41;
  - [arrayRandom](https://gitlab.com/201flaviosilva/utilsjs/-/commit/cbdb864e76c5b7d8379352f3524a6a1d85f171fb);
- - 
+
 ## [1.2.6] - 08-07-2022
  
 ### Added