+++ CHANGELOG.md
@@ -14,11 +14,14 @@ This file was based on [this template](https://gist.github.com/juampynr/4c18214a
 ## [1.2.9] - 29-07-2022
  
 ### Added
+- [Random Walk](https://gitlab.com/201flaviosilva/utilsjs/-/commit/56eebcaac033c05401140c4418b54e9579fb8d38) #46;
+- [Binary Search Tree](https://gitlab.com/201flaviosilva/utilsjs/-/commit/b141db4e7a0089a7c99a9ca8a70f7a37f73d59ef) #11;
 
 ### Changed
 - ["Skip" zones in range](https://gitlab.com/201flaviosilva/utilsjs/-/commit/b0d7a1d585f194c093ba22f621ebf71a0d5ce793) #68;
  
 ### Fixed
+- [Median](https://gitlab.com/201flaviosilva/utilsjs/-/commit/bb9e8d4aba6f64cc6752b7f0853460c2edc03d74);
 
 ### Breaking Changes
  - `findBigArrObj` and `getMaxArrayObjects` change to `arrayFindBigObject` #74;